apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
fastapi-users[sqlalchemy]
asyncpg
greenlet
python-dotenv
pyjwt
psycopg2-binary
resend
fastapi-users-db-sqlmodel
httpx_oauth
slowapi
